refactor: extract complex methods #37
